Replace omitted slash in feed URL generation
This slash was originally present, but I removed it at some point
because it was causing double-slashes. I believe the reason is that I
had a leading slash in my article URL pattern, which in retrospect
should not have been there. Omitting the slash caused problems for other
folks; I should have tested this better. This commit puts the slash back
where it belongs.