set a documentation_url in __init__, use it a log message; clean
Use official url for documentation. w/ computed version no, not 'notmyidea'.

The version of pelican is used to build the final url.  This may require that
the version number is bumped accordingly in dev branches, and I am not sure this
is a good idea (but it's mine ;)

While at it, in settings.py: reword warnings, cut long lines, reindent (flake8)