Ignore Emacs write-lock files.
Whenever a user edits a file, GNU Emacs creates symlinks prefixed with
.#, for example, ".#article.rst".  These symlinks are picked up by
Pelican when autoreload is enabled, causing many errors to be logged
until the edited file is saved in Emacs.  This change ignores files
starting with ".#" when determining whether the content has changed.

See
http://www.gnu.org/software/emacs/manual/html_node/emacs/Interlocking.html#Interlocking
for more on Emacs interlocking files.