test_generators: Fix ARCHIVES_SAVE_AS for ..._templates_save_as_false
This test started showing up as a failure after I fixed the mock
assertion from it's old:

  write.assert_called_count == 0

which was probably just adding another call to the `write` mock.  If
you put a proper test in place, like the current:

  self.assertFalse(bool(writer.mock_calls), writer.mock_calls)

The failure I'm fixing here dates back to at least the current master:

  commit be040715f984ceee69812c081b901653f8f08e08
  Author: Alexis Métaireau <alexis@notmyidea.org>
  Date:   Wed Jan 2 13:46:55 2013 +0100

    remove duplicates

and probably back to the original addition of this test.