Tag Cloud example code did not work
In the current version of Pelican, tag pages are generated at /tag/[keyword].html not /tag/[keyword]/

I had used this sample code and noted that the links were not rendering (giving me 404s). I looked at the template for "notmyidea" and noted that the links worked and were in a different format. I have not used older versions of pelican so I am not sure if previously this code was correct.