Update to the tag cloud example.
The example assumes that tag links will be [siteroot]/tag/[tag here]/, but the current version of pelican geenrate [siteroot]tag/[tag here].html

I noted that in the default notmyidea template the links were being generated differently and this resolved an issue I had with 404s. I also noted that if the keyword had spaces, you got %20 in the address field, but that with notmyidea it was giving with dashes ("-") and working. But I needed to run it through string first or there were errors.

I also noted in the taglist file, there is a completely different loop structure used for getting the tags, should that be used here instead? 
{% for tag in article.tags %}
<somehtml>{{ tag.url }}</somehtml>
{% endfor %}