Testing whether a reader class is enabled *before* calling it, otherwise the test is never reached.
For example, if Markdown hasn't been installed, and you attempt read_file('foo.md') then an exception is raised in MarkdownReader.__init__ as it tries to call Markdown, which is False.  The class field "enabled" was not checked before attempting to instantiate MarkdownReader.  Instead, there is a later test on the instance, which won't be reached.