Move the quickstart generation into a function
Quickstart files/dirs were generated with a short script at the end
of quickstart.main(), but this makes things really difficult to use
from a script because of the interactive dialog. Move the generation
of files into quickstart.quickstart() so that a script can just pass
in a config and get a quickstart directory.