Add `bare_title` to metadata when `title` exists
When `TYPOGRIFY=True` and you have a post with ALL CAPS in its title, the <title> tag of the page will incorrectly contain a <span class="caps"> element.

By allowing use of `bare_title` in templates, this allows you to use the "typogrified" version of the title in your page content, while using the bare title in the <title> tag.