Add period information to archive context.
When rendering the `archive.html` template using the year/month/day
archive feature, I cannot tell from the template whether this is a
alltime-, year-, month- or day-archive. I can use the date of the first
element inside the ``dates`` list, but the information is ambiguous, as
it's just a regular date object that could contain month and day
information even when rendering the year archive.

I hope the example templates make the issue more clear.