Clarify FAQ entry re: need to install Markdown
Folks keep running into this error, which probably signals a need to
change this behavior. After all, it wouldn't be hard for us to detect
what's going on and provide a better error message, such as: "It looks
like you're trying to process Markdown, but the Markdown library is not
currently installed. Please install the Python-Markdown library via 'pip
install markdown'."

Until we implement something akin to the above, this should serve as a
slightly-improved FAQ entry in the interim.