Add support for new article metadata modified date
Modified date is a standard feature in most blog engines.

Theme developer can use this date to show last updated time of an
article and enclose its value in `<time>` element.

We cannot use mtime of a file for modified_date because it is already
used for date metadata.

Secondly, a file mtime can be affected in a lot many ways, for example,
a sync service can change it, which will be incorrect to show as last
updated time of an article.