Add articles with subparts.
Sub-articles are useful when an article has logically related parts
that are too large to be conveniently included in a single article,
yet are not independent enough to merit a separate article.

The site pxquim.pt uses sub-articles to associate multiple photo galleries
with a single article. Each photo gallery goes into a sub-article.

By convention, subparts of articles are articles with "--" in their name.
The part before the "--" should be the slug of the parent article.

For example:

* article.md        (the parent article)
* article--one.md   (a sub-article of "article.md")
* article--other.md (another sub-article of "article.md")

The article object gets a list of subparts[].

The sub-article object gets the following attributes parent:

* parent -- Reference to the parent article object.
* title -- Changed to start with the parent article plus a comma.
* subtitle -- The original title of the sub-article.