Default --output to "content" instead of "output"
The default workflow of the rest of the tools expect the .rst|.md files to be in the content folder, but the pelican-import command was defaulting to placing those files in the output folder which is where pelican puts the generated html. This change creates a workflow which is more obvious for new users.