Ignore output for development server
The Makefile starts pelican to continuously generate the website and a
development web server both in the background, but the commands do still
output text to the shell. This behaviour is annoying when the same shell
is used further, therefore the output of both commands get redirected to
/dev/null.
Note: Even the -q, --quiet flag of pelican does not suppress
all output.

This fixes #1888