pelican-import, fix urls of attached files
This patch wp-attach modifies ```<img``` and  ```<a href``` urls to point at the attachment path - so  ```http://example.com/wp-content/uploads/blah.png```   becomes    ```wp-content/uploads/blah.png```

This needs wp-content in pelicanconf STATIC_FILES

I thought I could use {filename} but since the importer is generates source html this doesn't work.
It would be nice if there was a way to customise the output path, but not sure how that would work.