Show article category/tags even if no authors
Before this commit, the behavior of the simple theme is to hide the
category and tags if there is no author defined, even if the
category/tags are present. This appears to be unintentional, due to a
misplaced `endif`.

Fix this by moving the endif to the right place, so that the other if
blocks aren't nested.