Support --exclude-from for rsync publishing using excludes.txt
Add an excludes.txt example file where users can list rsync exclude
specifications.  If the site you are publishing has any paths on the
remote site that Pelican should ignore and not remove, list them in
excludes.txt so they are not removed each time you do:

    make rsync_upload

The example file has a comment pointing users to the rsync man page
for information on the format used in exclude files.  The only entry
in it by default is the .DS_Store entry which was from the rsync
command in tasks.py.