Improve logging of generators and writer loaders
This improves the logging of get_generator_classes() to
provide the source of the generator (either internal or from a plugin).
It also reviews the code to make it clearer (in my opinion, but I'm biased ;)), renames get_writer() to _get_writer_class() for consistency with the renamed _get_generator_classes().