Update Github actions badge
See the linked GitHub issue from the previous version of the badge: badges/shields#8671

In summary, these badges were changed in a breaking fashion such that the badges were just linking to the github issue rather than showing the data. Updating the url resolves this.

Also updated the link added to the badge image, so that it shows exactly the history of actions that the badge shows the most recent one of.