Add rel='nofollow' to all external hardcoded links in templates
Fix #2956: the HTML tag (eg `<a href="https://getpelican.com">`) in our
templates negatively impacts a Pelican-generated site's ranking for
Search Engine Optimisation purposes. This can be solved by adding the
Google-developed `rel="nofollow"` attribute to external links so that
the website's owner does not have to provide link authority to the
target. This change also had to be reflected in the tests so that they
would pass.